[battery_reminder] separate battery display header file
